+++ .gitignore
@@ -1,11 +1,22 @@
-/pack/
-/vendor/
+.idea
+.code
+**/vendor/
+.pulumi
 **/bin/
 **/obj/
-/node_modules/
-.pulumi
 Pulumi.*.yaml
-/.idea/
+**/node_modules/
+
+**/command-output/
+
+.idea/
 *.iml
+
 yarn.lock
 **/pulumiManifest.go
+
+ci-scripts
+**/schema.go
+**/version.txt
+**/nuget
+**/dist
