+++ Makefile.github
@@ -1,0 +1,96 @@
+PACK             := gcp
+ORG              := pulumi
+PROJECT          := github.com/${ORG}/pulumi-${PACK}
+NODE_MODULE_NAME := @pulumi/${PACK}
+TF_NAME          := ${PACK}
+PROVIDER_PATH    := provider/v2
+VERSION_PATH     := ${PROVIDER_PATH}/pkg/version.Version
+
+TFGEN           := pulumi-tfgen-${PACK}
+PROVIDER        := pulumi-resource-${PACK}
+VERSION         := $(shell pulumictl get version)
+
+LATEST_RESOURCE_PROVIDER_VERSION := $(shell pulumictl get latest-plugin pulumi-${PACK})
+PROVIDER_VERSION := ${LATEST_RESOURCE_PROVIDER_VERSION:v%=%}
+
+WORKING_DIR     := $(shell pwd)
+
+.PHONY: development provider build_sdks build_nodejs build_dotnet build_go build_python cleanup
+
+development:: install_plugins provider lint_provider build_sdks install_sdks cleanup # Build the provider & SDKs for a development environment
+
+tfgen:: install_plugins
+	(cd provider && go build -a -o $(WORKING_DIR)/bin/${TFGEN} -ldflags "-X ${PROJECT}/${VERSION_PATH}=${VERSION}" ${PROJECT}/${PROVIDER_PATH}/cmd/${TFGEN})
+	$(WORKING_DIR)/bin/${TFGEN} schema --out provider/cmd/${PROVIDER}
+	(cd provider && VERSION=$(VERSION) go generate cmd/${PROVIDER}/main.go)
+
+provider:: tfgen install_plugins # build the provider binary
+	(cd provider && go build -a -o $(WORKING_DIR)/bin/${PROVIDER} -ldflags "-X ${PROJECT}/${VERSION_PATH}=${VERSION}" ${PROJECT}/${PROVIDER_PATH}/cmd/${PROVIDER})
+
+build_sdks:: install_plugins provider build_nodejs build_python build_go build_dotnet # build all the sdks
+
+build_nodejs:: VERSION := $(shell pulumictl get version --language javascript)
+build_nodejs:: install_plugins tfgen # build the node sdk
+	$(WORKING_DIR)/bin/$(TFGEN) nodejs --overlays provider/overlays/nodejs --out sdk/nodejs/
+	cd sdk/nodejs/ && \
+        yarn install && \
+        yarn run tsc && \
+        cp ../../README.md ../../LICENSE package.json yarn.lock ./bin/ && \
+    	sed -i.bak -e "s/\$${VERSION}/$(VERSION)/g" ./bin/package.json
+
+build_python:: PYPI_VERSION := $(shell pulumictl get version --language python)
+build_python:: install_plugins tfgen # build the python sdk
+	$(WORKING_DIR)/bin/$(TFGEN) python --overlays provider/overlays/python --out sdk/python/
+	cd sdk/python/ && \
+        cp ../../README.md . && \
+        python3 setup.py clean --all 2>/dev/null && \
+        rm -rf ./bin/ ../python.bin/ && cp -R . ../python.bin && mv ../python.bin ./bin && \
+        sed -i.bak -e "s/\$${VERSION}/$(PYPI_VERSION)/g" -e "s/\$${PLUGIN_VERSION}/$(VERSION)/g" ./bin/setup.py && \
+        rm ./bin/setup.py.bak && \
+        cd ./bin && python3 setup.py build sdist
+
+build_go:: install_plugins tfgen # build the go sdk
+	$(WORKING_DIR)/bin/$(TFGEN) go --overlays provider/overlays/go --out sdk/go/
+
+build_dotnet:: DOTNET_VERSION := $(shell pulumictl get version --language dotnet)
+build_dotnet:: install_plugins tfgen # build the dotnet sdk
+	pulumictl get version --language dotnet
+	$(WORKING_DIR)/bin/$(TFGEN) dotnet --overlays provider/overlays/dotnet --out sdk/dotnet/
+	cd sdk/dotnet/ && \
+		echo "${DOTNET_VERSION}" >version.txt && \
+        dotnet build /p:Version=${DOTNET_VERSION}
+
+build_go:: install_plugins tfgen # build the go sdk
+	$(WORKING_DIR)/bin/$(TFGEN) go --overlays provider/overlays/go --out sdk/go/
+
+lint_provider:: provider # lint the provider code
+	cd provider && golangci-lint run -c ../.golangci.yml
+
+cleanup:: # cleans up the temporary directory
+	rm -r $(WORKING_DIR)/bin
+	rm -f provider/cmd/${PROVIDER}/schema.go
+
+help::
+	@grep '^[^.#]\+:\s\+.*#' Makefile | \
+ 	sed "s/\(.\+\):\s*\(.*\) #\s*\(.*\)/`printf "\033[93m"`\1`printf "\033[0m"`	\3 [\2]/" | \
+ 	expand -t20
+
+clean::
+	rm -rf sdk/{dotnet,nodejs,go,python}
+
+install_plugins::
+	[ -x $(shell which pulumi) ] || curl -fsSL https://get.pulumi.com | sh
+	pulumi plugin install resource $(PACK) $(PROVIDER_VERSION)
+
+install_dotnet_sdk::
+	mkdir -p $(WORKING_DIR)/nuget
+	find . -name '*.nupkg' -print -exec cp -p {} ${WORKING_DIR}/nuget \;
+
+install_python_sdk::
+
+install_go_sdk::
+
+install_nodejs_sdk::
+	yarn link --cwd $(WORKING_DIR)/sdk/nodejs/bin
+
+install_sdks:: install_dotnet_sdk install_python_sdk install_nodejs_sdk
