--- examples/bucket-py/.gitignore
@@ -1,2 +1,0 @@
-/.pulumi/
-*.pyc
