+++ examples/examples_test.go
@@ -18,8 +18,6 @@ import (
 	"io/ioutil"
 	"net/http"
 	"os"
-	"path"
-	"path/filepath"
 	"testing"
 	"time"
 
@@ -27,130 +25,6 @@ import (
 
 	"github.com/pulumi/pulumi/pkg/v2/testing/integration"
 )
-
-func TestAccMinimal(t *testing.T) {
-	test := getJSBaseOptions(t).
-		With(integration.ProgramTestOptions{
-			Dir: path.Join(getCwd(t), "minimal"),
-		})
-
-	integration.ProgramTest(t, &test)
-}
-
-func TestAccServerless(t *testing.T) {
-	test := getJSBaseOptions(t).
-		With(integration.ProgramTestOptions{
-			Dir:           path.Join(getCwd(t), "serverless"),
-			RunUpdateTest: true,
-			// One change is known to occur during refresh of the resources in this example:
-			// * `~  gcp:storage:Bucket f-bucket updated changes: + websites`
-			ExpectRefreshChanges: true,
-			ExtraRuntimeValidation: validateAPITest(func(body string) {
-				assert.Equal(t, "Hello World!", body)
-			}),
-		})
-
-	integration.ProgramTest(t, &test)
-}
-
-func TestAccBucket(t *testing.T) {
-	test := getJSBaseOptions(t).
-		With(integration.ProgramTestOptions{
-			Dir:           path.Join(getCwd(t), "bucket"),
-			RunUpdateTest: true,
-			// One change is known to occur during refresh of the resources in this example:
-			// * `~  gcp:storage:Bucket f-bucket updated changes: + websites`
-			ExpectRefreshChanges: true,
-		})
-
-	integration.ProgramTest(t, &test)
-}
-
-func TestAccTopic(t *testing.T) {
-	test := getJSBaseOptions(t).
-		With(integration.ProgramTestOptions{
-			Dir:           path.Join(getCwd(t), "topic"),
-			RunUpdateTest: true,
-			// One change is known to occur during refresh of the resources in this example:
-			// * `~  gcp:storage:Bucket f-bucket updated changes: + websites`
-			ExpectRefreshChanges: true,
-		})
-
-	integration.ProgramTest(t, &test)
-}
-
-func TestAccLoadbalancer(t *testing.T) {
-	skipIfShort(t)
-	test := getJSBaseOptions(t).
-		With(integration.ProgramTestOptions{
-			Dir: path.Join(getCwd(t), "loadbalancer"),
-			// TODO[pulumi/pulumi-terraform#241] This test currently triggers a bug in refresh, so we'll skip
-			// running the refresh step for now.
-			SkipRefresh:   true,
-			RunUpdateTest: true,
-			SkipPreview:   true,
-		})
-
-	integration.ProgramTest(t, &test)
-}
-
-func TestAccWebserver(t *testing.T) {
-	skipIfShort(t)
-	test := getJSBaseOptions(t).
-		With(integration.ProgramTestOptions{
-			Dir: path.Join(getCwd(t), "webserver"),
-			// TODO[pulumi/pulumi-terraform#241] This test currently triggers a bug in refresh, so we'll skip
-			// running the refresh step for now.
-			SkipRefresh:   true,
-			RunUpdateTest: true,
-		})
-
-	integration.ProgramTest(t, &test)
-}
-
-func TestAccMinimalPy(t *testing.T) {
-	test := getPythonBaseOptions(t).
-		With(integration.ProgramTestOptions{
-			Dir: path.Join(getCwd(t), "minimal-py"),
-		})
-
-	integration.ProgramTest(t, &test)
-}
-
-func TestAccBucketPy(t *testing.T) {
-	test := getPythonBaseOptions(t).
-		With(integration.ProgramTestOptions{
-			Dir:           path.Join(getCwd(t), "bucket-py"),
-			RunUpdateTest: true,
-		})
-
-	integration.ProgramTest(t, &test)
-}
-
-func TestAccDatasourcePy(t *testing.T) {
-	test := getPythonBaseOptions(t).
-		With(integration.ProgramTestOptions{
-			Dir:           path.Join(getCwd(t), "datasource-py"),
-			RunUpdateTest: true,
-		})
-
-	integration.ProgramTest(t, &test)
-}
-
-func TestAccAppServiceCs(t *testing.T) {
-	test := getCSBaseOptions(t).
-		With(integration.ProgramTestOptions{
-			Dir: path.Join(getCwd(t), "serverless-cs"),
-			// One change is known to occur during refresh of the resources in this example:
-			// BucketObject has a source change
-			ExpectRefreshChanges: true,
-			ExtraRuntimeValidation: validateAPITest(func(body string) {
-				assert.Equal(t, body, "Hello World!")
-			}),
-		})
-
-	integration.ProgramTest(t, &test)
-}
 
 func skipIfShort(t *testing.T) {
 	if testing.Short() {
@@ -205,39 +79,6 @@ func getBaseOptions(t *testing.T) integration.ProgramTestOptions {
 			"gcp:config:region":  region,
 		},
 	}
-}
-
-func getJSBaseOptions(t *testing.T) integration.ProgramTestOptions {
-	base := getBaseOptions(t)
-	baseJS := base.With(integration.ProgramTestOptions{
-		Dependencies: []string{
-			"@pulumi/gcp",
-		},
-	})
-
-	return baseJS
-}
-
-func getPythonBaseOptions(t *testing.T) integration.ProgramTestOptions {
-	base := getBaseOptions(t)
-	pythonBase := base.With(integration.ProgramTestOptions{
-		Dependencies: []string{
-			filepath.Join("..", "sdk", "python", "bin"),
-		},
-	})
-
-	return pythonBase
-}
-
-func getCSBaseOptions(t *testing.T) integration.ProgramTestOptions {
-	base := getBaseOptions(t)
-	csharpBase := base.With(integration.ProgramTestOptions{
-		Dependencies: []string{
-			"Pulumi.Gcp",
-		},
-	})
-
-	return csharpBase
 }
 
 func validateAPITest(isValid func(body string)) func(t *testing.T, stack integration.RuntimeValidationStackInfo) {
