--- examples/loadbalancer/.gitignore
@@ -1,4 +1,0 @@
-
-.pulumi/
-node_modules/
-bin
