--- examples/topic/.gitignore
@@ -1,3 +1,0 @@
-.pulumi/
-node_modules/
-bin
