--- provider/cmd/pulumi-resource-gcp/.gitignore
@@ -1,1 +1,0 @@
-schema.go
