--- sdk/.gitignore
@@ -1,3 +1,0 @@
-/nodejs/bin/
-/nodejs/node_modules/
-/python/bin/
