--- sdk/python/.gitignore
@@ -1,2 +1,0 @@
-# generated by the build
-README.rst
